Update project-demo.md
Scrum defines no demo - it is Review. It's goal is to gather feedback about increment (current state of the project) and the future state of the project (our plans for next sprint). Demo name suggests that it about only showing the product of sprint. It about gathering the feed back.